requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
absl-py==1.4.0
    # via
    #   -r requirements.in
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via soundfile
charset-normalizer==3.2.0
    # via requests
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
flatbuffers==23.5.26
    # via tensorflow
fonttools==4.42.0
    # via matplotlib
gast==0.4.0
    # via tensorflow
google-auth==2.22.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==1.0.0
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.56.2
    # via
    #   tensorboard
    #   tensorflow
h5py==3.9.0
    # via tensorflow
idna==3.4
    # via requests
jax==0.4.14
    # via tensorflow
joblib==1.3.1
    # via
    #   -r requirements.in
    #   scikit-learn
keras==2.12.0
    # via
    #   -r requirements.in
    #   tensorflow
kiwisolver==1.4.4
    # via matplotlib
libclang==16.0.6
    # via tensorflow
markdown==3.4.4
    # via tensorboard
markupsafe==2.1.3
    # via werkzeug
matplotlib==3.7.1
    # via
    #   -r requirements.in
    #   wfdb
ml-dtypes==0.2.0
    # via jax
numpy==1.23.5
    # via
    #   -r requirements.in
    #   contourpy
    #   h5py
    #   jax
    #   matplotlib
    #   ml-dtypes
    #   opt-einsum
    #   pandas
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow
    #   wfdb
oauthlib==3.2.2
    # via requests-oauthlib
opt-einsum==3.3.0
    # via
    #   jax
    #   tensorflow
packaging==23.1
    # via
    #   matplotlib
    #   tensorflow
pandas==2.0.2
    # via
    #   -r requirements.in
    #   wfdb
pillow==10.0.0
    # via matplotlib
protobuf==4.23.4
    # via
    #   tensorboard
    #   tensorflow
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2023.3
    # via pandas
pyyaml==6.0
    # via -r requirements.in
requests==2.31.0
    # via
    #   requests-oauthlib
    #   tensorboard
    #   wfdb
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
scikit-learn==1.3.0
    # via -r requirements.in
scipy==1.11.1
    # via
    #   -r requirements.in
    #   jax
    #   scikit-learn
    #   wfdb
six==1.16.0
    # via
    #   astunparse
    #   google-auth
    #   google-pasta
    #   python-dateutil
    #   tensorflow
soundfile==0.12.1
    # via wfdb
tensorboard==2.12.3
    # via tensorflow
tensorboard-data-server==0.7.1
    # via tensorboard
tensorflow==2.12.0
    # via -r requirements.in
tensorflow-estimator==2.12.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.33.0
    # via tensorflow
termcolor==2.3.0
    # via tensorflow
threadpoolctl==3.2.0
    # via scikit-learn
typing-extensions==4.7.1
    # via
    #   -r requirements.in
    #   tensorflow
tzdata==2023.3
    # via pandas
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
werkzeug==2.3.6
    # via tensorboard
wfdb==4.1.2
    # via -r requirements.in
wheel==0.41.0
    # via
    #   astunparse
    #   tensorboard
wrapt==1.14.1
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
